refactor sql_to_java
Use a giant switch statement instead of a series of if statements.
I think this makes it more clear that we are switching on the data
type.  It also gets rid of our JAVA_TYPE struct that was just a
confusing layer on top of the existing mysql_type struct (in my
opinion).